+++ cli/commands/metadata_diff.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/hasura/graphql-engine/cli/v2/internal/projectmetadata"
@@ -19,6 +20,11 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
+)
+
+const (
+	folderType string = "folder"
+	fileType   string = "file"
 )
 
 type MetadataDiffOptions struct {
@@ -51,7 +57,10 @@ By default, it shows changes between the exported metadata file and server metad
   hasura metadata diff local_metadata.yaml
 
   # Show changes between metadata from metadata.yaml and metadata_old.yaml:
-  hasura metadata diff metadata.yaml metadata_old.yaml
+  hasura metadata diff metadata.yaml ../old/metadata_old.yaml
+	
+  # Show changes among multiple metadata folders (v2 config only):
+  hasura metadata diff metadata1 metadata2
 
   # Apply admin secret for Hasura GraphQL engine:
   hasura metadata diff --admin-secret "<admin-secret>"
@@ -76,7 +85,7 @@ By default, it shows changes between the exported metadata file and server metad
 }
 
 func (o *MetadataDiffOptions) runv2(args []string) error {
-	messageFormat := "Showing diff between %s and %s..."
+	messageFormat := "Showing diff between %s directory and %s..."
 	message := ""
 	metadataHandler := projectmetadata.NewHandlerFromEC(o.EC)
 	from := "project"
@@ -87,21 +96,21 @@ func (o *MetadataDiffOptions) runv2(args []string) error {
 		from = "project"
 	case 1:
 		// 1 arg, diff given directory and the metadata on server
-		err := checkDir(args[0])
+		err := checkDirAndFile(args[0], folderType)
 		if err != nil {
 			return err
 		}
 		o.Metadata[0] = args[0]
 		from = o.Metadata[0]
 	case 2:
-		err := checkDir(args[0])
+		err := checkDirAndFile(args[0], folderType)
 		if err != nil {
 			return err
 		}
 		o.Metadata[0] = args[0]
 		from = o.Metadata[0]
 
-		err = checkDir(args[1])
+		err = checkDirAndFile(args[1], folderType)
 		if err != nil {
 			return err
 		}
@@ -215,23 +224,75 @@ func printDiffv1(before, after string, to io.Writer) {
 	for _, diff := range diffs {
 		text := diff.Payload
 		switch diff.Delta {
-		case difflib.RightOnly:
-			fmt.Fprintf(to, "%s\n", ansi.Color(text, "green"))
 		case difflib.LeftOnly:
-			fmt.Fprintf(to, "%s\n", ansi.Color(text, "red"))
-		case difflib.Common:
-			fmt.Fprintf(to, "%s\n", text)
+		case difflib.RightOnly:
+			isThereADiff = true
+			break
 		}
 	}
+
+	if isThereADiff {
+		logger.Info("displaying diff...")
+		fmt.Fprintf(to, "\n")
+		for _, diff := range diffs {
+			text := diff.Payload
+			// TODO: Probably printing just these is not enough,
+			// we may need to print 2-3 line before and after these
+			// changes to provide more context to the users
+			switch diff.Delta {
+			case difflib.RightOnly:
+				fmt.Fprintf(to, "%s\n", ansi.Color(text, "green"))
+			case difflib.LeftOnly:
+				fmt.Fprintf(to, "%s\n", ansi.Color(text, "red"))
+			}
+		}
+		return
+	}
+
+	logger.Info("no diff")
 }
 
-func checkDir(path string) error {
+// checkDirAndFile supports either "file" or "folder"
+func checkDirAndFile(path, pathType string) error {
 	file, err := os.Stat(path)
+
 	if err != nil {
+		if os.IsNotExist(err) {
+			return fmt.Errorf("this %s does not exist", pathType)
+		}
 		return err
 	}
-	if !file.IsDir() {
-		return fmt.Errorf("metadata diff only works with folder but got file %s", path)
+
+	if !file.IsDir() && pathType == folderType {
+		// This is meant for v2 config
+		return fmt.Errorf("metadata diff only works with metadata directories but got %s", path)
 	}
+
+	if file.IsDir() && pathType == fileType {
+		// This is meant for v1 config and no metadata.yml file wasn't found in the given dir.
+		return fmt.Errorf("metadata diff only works with metadata files but got %s", path)
+	}
+
 	return nil
+}
+
+func includeMetadataFileInPath(path string) string {
+	folder, err := os.Stat(path)
+	if err != nil {
+		return ""
+	}
+	if !folder.IsDir() {
+		return ""
+	}
+	files, err := ioutil.ReadDir(path)
+	if err != nil {
+		return ""
+	}
+	for _, file := range files {
+		currentFile := file.Name()
+		if currentFile == "metadata.yaml" || currentFile == "metadata.yml" {
+			return filepath.Join(path, currentFile)
+		}
+	}
+	return ""
 }